Fix parsing of virtual outputs
When disconnecting an output that is currently in use
xrandr leaves a virtual output as disconnected for the display.
The output of xrandr when this kind of display exists has the following format:

VIRTUAL1 disconnected (normal left inverted right x axis y axis)
  1920x1080 (0x4b) 148.500MHz
        h: width  1920 start 2008 end 2052 total 2200 skew    0 clock  67.50KHz
        v: height 1080 start 1084 end 1089 total 1125           clock  60.00Hz

The parser understands the last 3 lines as mode lines, and fails.
The fix ignores parsing the modes of disconnected displays, the results are now
consistent with the other type of disconnected outputs.